feat: add order by user

+++ src/infrastructure/repositories/dynamo/order.ts
@@ -1,4 +1,4 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
+ 
 import { docClient } from '../../../config/aws-config';
 import { Order } from '../../../domain/entities/order';
 import { OrderRepository } from '../../../domain/repositories/order';
@@ -10,24 +10,30 @@ export class DynamoOrderRepository implements OrderRepository {
     await docClient.put({ TableName: this.TABLE_NAME , Item: item }).promise();
     const result = await docClient.get({
       TableName: this.TABLE_NAME,
-      Key: { id: item.id },
+      Key: { id: item.id, customerId: item.customerId },
     }).promise();
     return result.Item as Order;
   }
 
   async delete(id: string, userId: string): Promise<void> {
-    await docClient.delete({ Key: { id }, TableName: this.TABLE_NAME }).promise();
+    await docClient.delete({ Key: { id, customerId: userId }, TableName: this.TABLE_NAME }).promise();
   }
 
   async findById(id: string, userId: string): Promise<Order | null> {
-    // ver pra buscar só as order q o user criou
-    const result = await docClient.get({ Key: { id }, TableName: this.TABLE_NAME }).promise();
+    const result = await docClient.get({ Key: { id, customerId: userId }, TableName: this.TABLE_NAME }).promise();
     return result.Item as Order;
   }
 
   async findAll(userId: string): Promise<Order[]> {
-    // ver pra buscar só as order q o user criou
-    const result = await docClient.scan({ TableName: this.TABLE_NAME }).promise();
+    const params = {
+      TableName: this.TABLE_NAME,
+      IndexName: 'customerId-index', 
+      KeyConditionExpression: 'customerId = :customerId',
+      ExpressionAttributeValues: {
+        ':customerId': userId,
+      },
+    };
+    const result = await docClient.query(params).promise();
     return result.Items as Order[];
   }
 }
